// src/models/IoK8sApiCoreV1ClusterTrustBundleProjection/types.ts
import { IoK8sApimachineryPkgApisMetaV1LabelSelectorModelOptions as IoK8sApimachineryPkgApisMetaV1LabelSelector } from '../IoK8sApimachineryPkgApisMetaV1LabelSelector';

export interface IoK8sApiCoreV1ClusterTrustBundleProjectionModelOptions {
  labelSelector?: IoK8sApimachineryPkgApisMetaV1LabelSelector;
  name?: string;
  optional?: boolean;
  path: string;
  signerName?: string;
}